refactor(web): rename main FilersPanel filename
