+++ apps/web/src/components/PaginatedTable/index.tsx
@@ -12,7 +12,7 @@ import { Pagination } from "~/components/Pagination";
 import type { TableProps } from "~/components/Table";
 import { Table } from "~/components/Table";
 import type { Rollup } from "~/types";
-import { FilterPanel } from "./components/FilterPanel";
+import { FilterPanel } from "./components/Filters";
 
 const DEFAULT_TABLE_EMPTY_STATE = "No items";
 const PAGE_SIZES_OPTIONS: DropdownProps["options"] = [
